now you can download game

+++ src/main/kotlin/tasks/DownloadGame.kt
@@ -15,18 +15,40 @@ object DownloadGame: Callable<Int> {
 
     override fun call(): Int {
         logger.info("Reading project...")
-        val project = Project.read().project
+        if (Project.read() == null) return 1
+        val project = Project.read()!!.project
+
+        logger.info("Getting versions")
+        val gameVersions = project.platform.realisation.getGameVersions()
+        val loaderVersions = project.platform.realisation.getLoaderVersions()
+
+        if (project.minecraft !in gameVersions) {
+            logger.error("Minecraft ${project.minecraft} not found")
+            return 1
+        }
+        if (project.loader !in loaderVersions) {
+            logger.error("Loader ${project.loader} not found")
+            return 1
+        }
 
         logger.info("Clear old game")
-        val versionsFolder = File(Project.getGameFolder(), "versions")
+        val versionsFolder = File(Project.gameFolder, "versions")
         versionsFolder.deleteRecursively()
         versionsFolder.mkdirs()
-        Project.getGameVersionFile().delete()
+        val librariesFolder = File(Project.gameFolder, "libraries")
+        librariesFolder.deleteRecursively()
+        librariesFolder.mkdirs()
 
-        logger.info("Getting game versions")
-        val versions = project.platform.realisation.getGameVersions()
-        logger.info(versions.toString())
+        logger.info("Starting installation...")
+        val result = project.platform.realisation.install(project.minecraft, project.loader, Project.gameFolder)
 
-        return 0
+        if (result == 0) {
+            logger.info("Your game successfully installed!")
+            return 0
+        }
+        else {
+            logger.error("Error installing game $result")
+            return result
+        }
     }
 }
